backend/requirements.txt
# Core
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
python-multipart
sqlalchemy
asyncpg
alembic
aiofiles
aiosqlite

# PDF & Image Processing
pdf2image
pillow
opencv-python-headless

# AI Models (Local)
transformers
torch
torchvision
accelerate
bitsandbytes
tiktoken
huggingface-hub

# AI Models (Gemini)
google-genai

# Utilities
requests
tqdm
numpy
pandas
